requirements.txt
PyGithub
python-dotenv
selenium